# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Application API",
        "description": "API documentation for My Application",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "description": "Создаёт нового пользователя и возвращает API-токен (Sanctum)",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "token"
                                    ],
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|randomlongtokenhere1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Логин пользователя",
                "description": "Логин пользователя и возвращение API-токена (Sanctum)",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешный логин",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "token"
                                    ],
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|randomlongtokenhere1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Логаут пользователя",
                "description": "Логаут пользователя и удаление API-токена (Sanctum)",
                "operationId": "logout",
                "responses": {
                    "204": {
                        "description": "Успешный логаут"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@mail.ru",
                        "maxLength": 255
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!",
                        "minLength": 8
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "surname",
                    "email",
                    "phone",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Иван",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "surname": {
                        "type": "string",
                        "example": "Иванов",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "patronymic": {
                        "type": "string",
                        "example": "Иванович",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com",
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "example": "+79991234567"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}